Update NFT.sol
Added NATSPEC comments

+++ contracts/NFT.sol
@@ -1,52 +1,61 @@
-// SPDX-License-Identifier: UNLICENSED
+// SPDX-License-Identifier: MIT License
 pragma solidity ^0.8.4;
 
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "erc721a/contracts/ERC721A.sol";
 
+/// @title UCL Blockchain Labs NFT
+/// @author UCL Blockchain Labs Division
+/// @notice ERC721 NFT with additional airdrop functionality
+/// @inheritdoc ERC721A
 contract NFT is ERC721A, Ownable {
-    // TODO: ADD COMMENTS
-    // TODO: change names (contract and tokens)
-    uint public immutable maxSupply;
+    uint256 public immutable maxSupply;
     string public baseURI;
     string public contractURI;
 
-    event ContractURIChanged(
-        address sender,
-        string newContractURI
-    );
+    event ContractURIChanged(address sender, string newContractURI);
 
-    event BaseURIChanged (
-        address sender,
-        string newBaseURI
-    );
+    event BaseURIChanged(address sender, string newBaseURI);
 
-    event Airdropped (
-        address recipient,
-        address sender,
-        uint amount
-    );
+    event Airdropped(address recipient, address sender, uint256 amount);
 
-    constructor(uint _maxSupply, string memory _baseURI, string memory _contractURI) ERC721A("NFT_TESTING_1", "NFT1") {
+    constructor(
+        uint256 _maxSupply,
+        string memory _baseURI,
+        string memory _contractURI
+    ) ERC721A("NFT_TESTING_1", "NFT1") {
         maxSupply = _maxSupply;
         baseURI = _baseURI;
         contractURI = _contractURI;
     }
 
+    /// @notice Update baseURI of NFT
+    /// @param _newBaseURI New baseURI
     function setBaseURI(string calldata _newBaseURI) public onlyOwner {
         baseURI = _newBaseURI;
         emit BaseURIChanged(msg.sender, _newBaseURI);
     }
 
+    /// @notice Update contractURI/NFT metadata
+    /// @param _newContractURI New metadata
     function setContractURI(string calldata _newContractURI) public onlyOwner {
         contractURI = _newContractURI;
         emit ContractURIChanged(msg.sender, _newContractURI);
     }
 
-    function airdrop(address[] calldata _recipients) external payable onlyOwner {
+    /// @notice Airdrop nft to given addresses
+    /// @param _recipients Array of addresses to airdrop nft to
+    function airdrop(address[] calldata _recipients)
+        external
+        payable
+        onlyOwner
+    {
         // `_mint`'s second argument now takes in a `quantity`, not a `tokenId`.
-        require(totalSupply() + _recipients.length <= maxSupply, "Max supply reached");
-        for (uint i = 0; i < _recipients.length; i++) {
+        require(
+            totalSupply() + _recipients.length <= maxSupply,
+            "Max supply reached"
+        );
+        for (uint256 i = 0; i < _recipients.length; i++) {
             address to = _recipients[i];
             require(to != address(0), "Can't mint to zero address");
             _safeMint(to, 1);
@@ -54,28 +63,29 @@ contract NFT is ERC721A, Ownable {
         }
     }
 
+    /// @notice Return baseURI
     function _baseURI() internal view virtual override returns (string memory) {
         return baseURI;
     }
 
     function tokenURI(uint256 tokenId)
-    public
-    view
-    virtual
-    override
-    returns (string memory)
+        public
+        view
+        virtual
+        override
+        returns (string memory)
     {
         string memory uri = super.tokenURI(tokenId);
         return
-        bytes(uri).length > 0
-        ? string(abi.encodePacked(uri, ".json"))
-        : "";
+            bytes(uri).length > 0 ? string(abi.encodePacked(uri, ".json")) : "";
     }
 
+    /// @return _ownershipOf Return address of owner of NFT with given tokenID
+    /// @param tokenId The token id to check ownership of.
     function getOwnershipData(uint256 tokenId)
-    external
-    view
-    returns (TokenOwnership memory)
+        external
+        view
+        returns (TokenOwnership memory)
     {
         return _ownershipOf(tokenId);
     }
